+++ app/cells/folio/console/atoms/previews/previews.js
@@ -1,2 +1,2 @@
-# see app/assets/stylesheets/folio/console/atoms/previews/base.sass
-# no code here as it would get included by a wildcard import
+// see app/assets/stylesheets/folio/console/atoms/previews/base.coffee
+// no code here as it would get included by a wildcard import